% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualizations.R
\name{viz_01_poverty_scatterplot}
\alias{viz_01_poverty_scatterplot}
\title{Viz 01 Porverty Scatterplot}
\usage{
viz_01_poverty_scatterplot(data)
}
\arguments{
\item{data}{data.frame. Made for use with [upwork03::averages]}
}
\description{
Viz 01 Porverty Scatterplot
}
