% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualizations.R
\name{viz_02_correlation_heatmap}
\alias{viz_02_correlation_heatmap}
\title{Viz 02 Correlation Heatmap}
\usage{
viz_02_correlation_heatmap(
  year,
  country,
  fontsize = 16,
  vars = c("Poverty Ratio", "Health Expenditure", "Drinking Water", "Sanitation")
)
}
\arguments{
\item{year}{numeric.}

\item{country}{char vector.}

\item{fontsize}{numeric.}

\item{vars}{vars to be showed on correlation heatmap}
}
\description{
Viz 02 Correlation Heatmap
}
