% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualizations.R
\name{viz_03_poverty_ratio}
\alias{viz_03_poverty_ratio}
\title{Viz 03 Poverty Ratio}
\usage{
viz_03_poverty_ratio(year, country, fontsize = 16, ylab = "Poverty ratio")
}
\arguments{
\item{year}{numeric.}

\item{fontsize}{numeric.}
}
\description{
Viz 03 Poverty Ratio
}
