% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualizations.R
\name{viz_04_highest_incidence}
\alias{viz_04_highest_incidence}
\title{Viz 04 Highest Incidence}
\usage{
viz_04_highest_incidence(year, TMH, country, fontsize = 16)
}
\arguments{
\item{year}{numeric.}

\item{TMH}{character.}

\item{fontsize}{numeric.}
}
\description{
Viz 04 Highest Incidence
}
