% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualizations.R
\name{viz_05_incidence_growth}
\alias{viz_05_incidence_growth}
\title{Viz 05 Poverty Ratio}
\usage{
viz_05_incidence_growth(country, fontsize = 16)
}
\arguments{
\item{fontsize}{numeric.}
}
\description{
Viz 05 Poverty Ratio
}
